+++ app/src/main/kotlin/com/makentoshe/booruchan/booru/model/gallery/posts/PostOrderedInfinityGallery.kt
@@ -1,4 +1,4 @@
-package com.makentoshe.booruchan.booru.model.gallery.post.ordinf
+package com.makentoshe.booruchan.booru.model.gallery.posts
 
 import android.support.v7.widget.LinearLayoutManager
 import android.view.View
